修复功能

+++ src/main/java/zgz/community2/controller/publishController.java
@@ -31,7 +31,7 @@ public class publishController {
 
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -52,7 +52,7 @@ public class publishController {
     public String doPublish(@RequestParam(value = "title",required = false) String title,
                             @RequestParam(value = "description") String description,
                             @RequestParam(value = "tag",required = false) String tag,
-                            @RequestParam(value = "id",required = false) Integer id,
+                            @RequestParam(value = "id",required = false) Long id,
                             HttpServletRequest request,
                             Model model){
 
@@ -64,7 +64,6 @@ public class publishController {
         User user = (User) request.getSession().getAttribute("user");
 
         User user1 = userMapper.userName(user.getUsername());
-
 
         Question question = new Question();
         question.setTitle(title);
